Implement ResumePage component for résumé analysis with file upload, loading states, and extraction results
