+++ packages/web/src/App.tsx
@@ -1,70 +1,17 @@
-import { useState } from 'react';
-import type { ResumeExtraction } from '@resume-lens/shared';
-import { ResumeLensError } from '@resume-lens/shared';
+import { JSX } from 'react';
 
-import { extractResume, NetworkError } from '@/api/client';
-import { getErrorMessage, NETWORK_ERROR_MESSAGE } from '@/utils/error-messages';
+import ResumePage from './pages/ResumePage';
 import { ThemeProvider } from '@/contexts/ThemeContext';
-import Header from '@/components/Header';
-import UploadForm from '@/components/UploadForm';
-import LoadingState from '@/components/LoadingState';
-import ResultCard from '@/components/ResultCard';
+import { TooltipProvider } from '@/components/shadcn/tooltip';
 
 import '@/index.css';
 
-const App = () => {
-  const [isLoading, setIsLoading] = useState(false);
-  const [extractionResult, setExtractionResult] = useState<ResumeExtraction | null>(null);
-  const [error, setError] = useState<string | null>(null);
-
-  const handleUpload = async (file: File) => {
-    try {
-      const result = await extractResume(file);
-      setExtractionResult(result);
-    } catch (err) {
-      let errorMessage = 'An unexpected error occurred';
-
-      if (err instanceof ResumeLensError) {
-        errorMessage = getErrorMessage(err.code);
-      } else if (err instanceof NetworkError) {
-        errorMessage = NETWORK_ERROR_MESSAGE;
-      } else if (err instanceof Error) {
-        errorMessage = err.message;
-      }
-
-      setError(errorMessage);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  /**
-   * Called when user starts form submission.
-   * Clears any existing errors so we have a clean state for the new request.
-   */
-  const handleStartSubmit = () => {
-    setError(null);
-    setIsLoading(true);
-  };
-
-  /**
-   * Resets the app state to allow analyzing another résumé.
-   */
-  const handleReset = () => {
-    setExtractionResult(null);
-    setError(null);
-  };
-
+const App = (): JSX.Element => {
   return (
     <ThemeProvider defaultTheme="dark" storageKey="resume-lens-ui-theme">
-      <div>
-        <Header />
-        {isLoading && <LoadingState />}
-        {!isLoading && !extractionResult && (
-          <UploadForm onUpload={handleUpload} onStartSubmit={handleStartSubmit} error={error} />
-        )}
-        {extractionResult && <ResultCard extraction={extractionResult} onReset={handleReset} />}
-      </div>
+      <TooltipProvider>
+        <ResumePage />
+      </TooltipProvider>
     </ThemeProvider>
   );
 };
